sample/fhir-producer/requirements.txt
kafka-python==2.0.2
fhir.resources==7.0.2